Fix: Only call setOriginalImageSize() on single page images
