+++ src/lib/viewers/image/ImageViewer.js
@@ -3,7 +3,7 @@ import Browser from '../../Browser';
 import ImageBaseViewer from './ImageBaseViewer';
 import { ICON_FILE_IMAGE, ICON_FULLSCREEN_IN, ICON_FULLSCREEN_OUT, ICON_ROTATE_LEFT } from '../../icons/icons';
 import { CLASS_INVISIBLE } from '../../constants';
-import { openContentInsideIframe } from '../../util';
+import { get, openContentInsideIframe } from '../../util';
 import './Image.scss';
 
 const CSS_CLASS_IMAGE = 'bp-image';
@@ -31,6 +31,21 @@ class ImageViewer extends ImageBaseViewer {
         this.imageEl.classList.add(CLASS_INVISIBLE);
 
         this.currentRotationAngle = 0;
+    }
+
+    /**
+     * @inheritdoc
+     *
+     * @return {void}
+     */
+    finishLoading() {
+        if (this.isDestroyed()) {
+            return;
+        }
+
+        this.setOriginalImageSize(this.imageEl)
+            .then(() => super.finishLoading())
+            .catch(this.errorHandler);
     }
 
     /**
@@ -99,7 +114,7 @@ class ImageViewer extends ImageBaseViewer {
      * @return {void}
      */
     rotateLeft() {
-        this.currentRotationAngle = (this.currentRotationAngle - 90) % 3600 % 360;
+        this.currentRotationAngle = ((this.currentRotationAngle - 90) % 3600) % 360;
         this.imageEl.setAttribute('data-rotation-angle', this.currentRotationAngle);
         this.imageEl.style.transform = `rotate(${this.currentRotationAngle}deg)`;
         this.emit('rotate');
@@ -238,7 +253,7 @@ class ImageViewer extends ImageBaseViewer {
         this.scale = width
             ? width / this.imageEl.getAttribute('originalWidth')
             : height / this.imageEl.getAttribute('originalHeight');
-        this.rotationAngle = this.currentRotationAngle % 3600 % 360;
+        this.rotationAngle = (this.currentRotationAngle % 3600) % 360;
         this.emit('scale', {
             scale: this.scale,
             rotationAngle: this.rotationAngle
@@ -397,11 +412,58 @@ class ImageViewer extends ImageBaseViewer {
         this.adjustImageZoomPadding();
 
         this.scale = this.imageEl.clientWidth / this.imageEl.getAttribute('originalWidth');
-        this.rotationAngle = this.currentRotationAngle % 3600 % 360;
+        this.rotationAngle = (this.currentRotationAngle % 3600) % 360;
         this.emit('scale', {
             scale: this.scale,
             rotationAngle: this.rotationAngle
         });
+    }
+
+    /**
+     * Sets the original image width and height on the img element. Can be removed when
+     * naturalHeight and naturalWidth attributes work correctly in IE 11.
+     *
+     * @private
+     * @param {HTMLElement} imageEl - The image to set the original size attributes on
+     * @return {Promise} A promise that is resolved if the original image dimensions were set.
+     */
+    setOriginalImageSize(imageEl) {
+        const promise = new Promise((resolve) => {
+            // Do not bother loading a new image when the natural size attributes exist
+            if (imageEl.naturalWidth && imageEl.naturalHeight) {
+                imageEl.setAttribute('originalWidth', imageEl.naturalWidth);
+                imageEl.setAttribute('originalHeight', imageEl.naturalHeight);
+                resolve();
+            } else {
+                // Case when natural dimensions are not assigned
+                // By default, assigned width and height in Chrome/Safari/Firefox will be 300x150.
+                // IE11 workaround. Dimensions only displayed if the image is attached to the document.
+                get(imageEl.src, {}, 'text')
+                    .then((imageAsText) => {
+                        const parser = new DOMParser();
+                        const svgEl = parser.parseFromString(imageAsText, 'image/svg+xml');
+
+                        try {
+                            // Assume svgEl is an instanceof an SVG with a viewBox and preserveAspectRatio of meet
+                            // where the height is the limiting axis
+                            const viewBox = svgEl.documentElement.getAttribute('viewBox');
+                            const [, , w, h] = viewBox.split(' ');
+                            const aspectRatio = h ? w / h : w;
+                            imageEl.setAttribute('originalWidth', Math.round(aspectRatio * 150));
+                            imageEl.setAttribute('originalHeight', 150);
+                        } catch (e) {
+                            // Assume 300x150 that chrome does by default
+                            imageEl.setAttribute('originalWidth', 300);
+                            imageEl.setAttribute('originalHeight', 150);
+                        } finally {
+                            resolve();
+                        }
+                    })
+                    .catch(resolve);
+            }
+        });
+
+        return promise;
     }
 }
